Moved function to helpers.ts
